chore: decouple memoized as separated module
